% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FizzBuzz.R
\name{FizzBuzz}
\alias{FizzBuzz}
\title{Fizz Buzz Function}
\usage{
FizzBuzz(n)
}
\arguments{
\item{n}{: a positive integer}
}
\description{
Because this is a very simple function, my explanation is short. These
paragraphs should explain everything you need to know.
}
\details{
This is a function that will output a  vector of the
first $n$ terms in the child's game *Fizz Buzz*. The input of the function is
a positive integer $n$ and the output will be a vector of characters. The goal
is to count as high as you can, but for any number evenly divisible by 3,
substitute "Fizz" and any number evenly divisible by 5, substitute "Buzz",
and if it is divisible by both, substitute "Fizz Buzz". So the sequence will
look like 1, 2, Fizz, 4, Buzz, Fizz, 7, 8, Fizz, ..
}
